jupyters/ticket_classif_ex2/deps/pip_requirements.txt
absl-py==0.13.0
alabaster==0.7.12
anyio==2.2.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
asn1crypto==1.4.0
astor==0.8.1
astroid==2.5
astropy==4.2.1
astunparse==1.6.3
async_generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.6
babel==2.9.0
backcall==0.2.0
backports==1.0
backports.functools_lru_cache==1.6.4
backports.shutil_get_terminal_size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bitarray==1.9.2
bkcharts==0.2
black==19.10b0
blas==1.0
bleach==3.3.0
blinker==1.4
blosc==1.21.0
bokeh==2.3.2
boto==2.49.0
bottleneck==1.3.2
brotli==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2021.10.8
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
charls==2.2.0
click==7.1.2
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.4
comtypes==1.1.9
conda==4.11.0
conda-build==3.21.4
conda-content-trust==0.1.1
conda-env==2.6.0
conda-package-handling==1.7.3
conda-repo-cli==1.0.4
conda-token==0.3.0
conda-verify==3.4.2
console_shortcut==0.1.1
contextlib2==0.6.0.post1
coverage==5.5
cryptography==3.4.7
curl==7.71.1
cycler==0.10.0
cython==0.29.23
cytoolz==0.11.0
dask==2021.4.0
dask-core==2021.4.0
decorator==5.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distributed==2021.4.0
docutils==0.17
entrypoints==0.3
et_xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.9.0
flask==1.1.2
freetype==2.10.4
fsspec==0.9.0
future==0.18.2
gast==0.4.0
get_terminal_size==1.0.0
gevent==21.1.2
giflib==5.2.1
glob2==0.7
gmpy2==2.1.0rc1
google-auth==1.21.3
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==2.38
greenlet==1.0.0
grpcio==1.36.1
h5py==2.10.0
hdf5==1.10.4
heapdict==1.0.1
html5lib==1.1
icc_rt==2019.0.0
icu==58.2
idna==2.10
imagecodecs==2021.3.31
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.10.0
importlib_metadata==3.10.0
iniconfig==1.1.1
intel-openmp==2021.2.0
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.22.0
ipython_genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
jinja2==2.11.3
joblib==1.0.1
jpeg==9b
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-packaging==0.7.12
jupyter_client==6.1.12
jupyter_console==6.4.0
jupyter_core==4.7.1
jupyter_server==1.4.1
jupyterlab==3.0.14
jupyterlab_pygments==0.1.2
jupyterlab_server==2.4.0
jupyterlab_widgets==1.0.0
keras==2.4.3
keras-applications==1.0.8
keras-base==2.4.3
keras-preprocessing==1.1.2
keyring==22.3.0
kiwisolver==1.3.1
krb5==1.18.2
lazy-object-proxy==1.6.0
lcms2==2.12
lerc==2.2.1
libaec==1.0.4
libarchive==3.4.2
libcurl==7.71.1
libdeflate==1.7
libiconv==1.15
liblief==0.10.1
libllvm9==9.0.1
libpng==1.6.37
libprotobuf==3.17.2
libsodium==1.0.18
libspatialindex==1.9.3
libssh2==1.9.0
libtiff==4.2.0
libxml2==2.9.10
libxslt==1.1.34
libzopfli==1.0.3
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
lz4-c==1.9.3
lzo==2.10
m2w64-gcc-libgfortran==5.3.0
m2w64-gcc-libs==5.3.0
m2w64-gcc-libs-core==5.3.0
m2w64-gmp==6.1.0
m2w64-libwinpthread-git==5.0.0.4634.697f757
markdown==3.3.4
markupsafe==1.1.1
matplotlib==3.3.4
matplotlib-base==3.3.4
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl==2021.2.0
mkl-service==2.3.0
mkl_fft==1.3.0
mkl_random==1.2.1
mock==4.0.3
more-itertools==8.7.0
mpc==1.2.1
mpfr==4.1.0
mpir==3.0.0
mpmath==1.2.1
msgpack-python==1.0.2
msys2-conda-epoch==20160418
multipledispatch==0.6.0
mypy_extensions==0.4.3
navigator-updater==0.2.1
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.6.1
nose==1.3.7
notebook==6.3.0
numba==0.53.1
numexpr==2.7.3
numpy==1.20.1
numpy-base==1.20.1
numpydoc==1.1.0
oauthlib==3.1.1
olefile==0.46
openjpeg==2.3.0
openpyxl==3.0.7
openssl==1.1.1l
opt_einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandoc==2.12
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
partd==1.2.0
path==15.1.2
path.py==12.5.0
pathlib2==2.3.5
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.2.0
pip==21.0.1
pkginfo==1.7.0
plotly==5.1.0
pluggy==0.13.1
ply==3.11
powershell_shortcut==0.0.1
prometheus_client==0.10.1
prompt-toolkit==3.0.17
prompt_toolkit==3.0.17
protobuf==3.17.2
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-lief==0.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.6
pydocstyle==6.0.0
pydotplus==2.0.2
pyerfa==1.7.3
pyflakes==2.2.0
pygments==2.8.1
pyjwt==2.1.0
pylint==2.7.4
pyls-black==0.4.6
pyls-spyder==0.3.2
pynacl==1.4.0
pyodbc==4.0.30
pyopenssl==20.0.1
pyparsing==2.4.7
pyqt==5.9.2
pyreadline==2.1
pyrsistent==0.17.3
pysocks==1.7.1
pytables==3.6.1
pytest==6.2.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-libarchive-c==2.9
python_abi==3.8
pytz==2021.1
pywavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyyaml==5.4.1
pyzmq==20.0.0
qdarkstyle==2.8.1
qt==5.9.7
qtawesome==1.0.2
qtconsole==5.0.3
qtpy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.2
rtree==0.9.7
ruamel_yaml==0.15.100
scikit-image==0.18.1
scipy==1.6.2
seaborn==0.11.1
send2trash==1.5.0
setuptools==52.0.0
simplegeneric==0.8.1
singledispatch==3.6.1
sip==4.19.13
six==1.15.0
sklearn-crfsuite==0.3.6
snappy==1.1.8
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcollections==2.1.0
sortedcontainers==2.3.0
soupsieve==2.2.1
sphinx==4.0.1
sphinxcontrib==1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
spyder==4.2.5
spyder-kernels==1.10.2
sqlalchemy==1.4.7
sqlite==3.35.4
statsmodels==0.12.2
sympy==1.8
tabulate==0.8.9
tbb==2020.3
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-base==2.3.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
textdistance==4.2.1
threadpoolctl==2.1.0
three-merge==0.1.1
tifffile==2021.4.8
tk==8.6.10
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.2
typing_extensions==3.7.4.3
ujson==4.0.2
unicodecsv==0.14.1
urllib3==1.26.4
vc==14.2
vs2015_runtime==14.27.29016
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
win_inet_pton==1.1.0
win_unicode_console==0.5
wincertstore==0.2
winpty==0.4.3
wordcloud==1.6.0
wrapt==1.12.1
xlrd==2.0.1
xlsxwriter==1.3.8
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
xz==5.2.5
yaml==0.2.5
yapf==0.31.0
zeromq==4.3.3
zfp==0.5.5
zict==2.0.0
zipp==3.4.1
zlib==1.2.11
zope==1.0
zope.event==4.5.0
zope.interface==5.3.0
zstd==1.4.5
blis==0.7.5
catalogue==2.0.6
catboost==1.0.3
cymem==2.0.6
en-core-web-sm==3.2.0
gensim==4.1.2
jovian==0.2.38
kmodes==0.11.0
langcodes==3.3.0
murmurhash==1.0.6
pathy==0.6.1
preshed==3.0.6
pycrf==0.0.1
pydantic==1.8.2
python-graphviz==0.17
qdldl==0.1.5.post0
scikit-learn==0.22.2
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
tensor-sensor==1.0
textblob==0.17.1
thinc==8.0.13
typer==0.4.0
uuid==1.30
wasabi==0.8.2
xgboost==1.4.2
